Added tag attribute to ResourceNode
ResourceNodes may be tagged and the tag can be retrieved via HTTPRequest::getTag(),
which allows grouping similar nodes and processing them in the same way. For example
if you want to have access protection for some nodes, give them a "PROTECTED" tag and
check for it in some authorization middleware function.